updated feedback controllers

+++ project/libs/feedback-lib/src/feedback.service.ts
@@ -1,7 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateCommentDto } from './dtos';
 import { FeedbackRepository } from './feedback.repository';
 import { FeedbackEntity } from './feedback.entity';
+import { ErrorMessages } from '@project/core';
 
 @Injectable()
 export class FeedbackService {
@@ -17,7 +22,13 @@ export class FeedbackService {
     return this.postRepository.findByPostId(postId);
   }
 
-  public async delete(id: string) {
-    return this.postRepository.deleteById(id);
+  public async delete(userId: string, id: string) {
+    const comment = await this.postRepository.findById(id);
+    if (comment) {
+      if (comment.createdBy === userId)
+        return this.postRepository.deleteById(id);
+      throw new BadRequestException(ErrorMessages.CommentUserError);
+    }
+    throw new NotFoundException(ErrorMessages.CommentNotFound);
   }
 }
